Renamed sighting list page folder
