+++ frontend/src/components/unapprovedSightings/UnapprovedSightingPage.tsx
